+++ Src/ITWeiHan.HtmlTableHelper.Test/ITWeiHan.HtmlTableHelperTests.cs
@@ -1,7 +1,9 @@
+using Dapper;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
 using System.Collections.Generic;
 using System.Data;
+using System.Data.SQLite;
 using System.Linq;
 
 namespace HtmlTableHelper.Test
@@ -43,19 +45,48 @@ namespace HtmlTableHelper.Test
         }
 
         [TestMethod]
+        public void Cache_Test()
+        {
+            for (int i = 0; i < 2; i++)
+            {
+                IEnumerable_Test();
+            }
+        }
+
+        [TestMethod]
+        public void DapperDynamicQuery_Test()
+        {
+            var expected = @"<table><thead><tr><th>Name</th><th>Age</th><th>Country</th></tr></thead><tbody><tr><td>ITWeiHan</td><td>25</td><td>Taiwan</td></tr></tbody></table>";
+            const string _path = "Test.sqlite";
+            SQLiteConnection.CreateFile(_path);
+            using (var cn = new SQLiteConnection($"Data Source={_path};Version=3;"))
+            {
+                var sourceData = cn.Query(@"select 'ITWeiHan' Name,25 Age,'Taiwan' Country");
+                var html = sourceData.ToHtmlTable();
+                Assert.AreEqual(html, expected);
+            }
+        }
+
+        [TestMethod]
         public void DataTableToHtml_Test()
         {
             var expected = @"<table><thead><tr><th>Name</th><th>Age</th><th>Country</th></tr></thead><tbody><tr><td>ITWeiHan</td><td>25</td><td>Taiwan</td></tr></tbody></table>";
 
+            DataTable table = GetTestDataTable();
+
+            var html = table.ToHtmlTable();
+            Assert.AreEqual(html, expected);
+        }
+
+        private static DataTable GetTestDataTable()
+        {
             var table = new DataTable();
             table.Columns.Add("Name", typeof(string));
             table.Columns.Add("Age", typeof(int));
             table.Columns.Add("Country", typeof(string));
 
             table.Rows.Add("ITWeiHan", 25, "Taiwan");
-
-            var html = table.ToHtmlTable();
-            Assert.AreEqual(html, expected);
+            return table;
         }
 
         [TestMethod]
@@ -86,7 +117,7 @@ namespace HtmlTableHelper.Test
         public void NonEncodeModeTest()
         {
             var expected = @"<table><thead><tr><th>Name</th></tr></thead><tbody><tr><td><b>ITWeiHan</b></td></tr></tbody></table>";
-            var htmltablesetting = new HTMLTableSetting()
+            var htmltablesetting = new HtmlTableSetting()
             {
                 IsHtmlEncodeMode = false
             };
@@ -99,30 +130,64 @@ namespace HtmlTableHelper.Test
         public void AttributeTableTest()
         {
             var expected = @"<table class=""SomeClass"" ><thead><tr><th>Name</th><th>Age</th><th>Country</th></tr></thead><tbody><tr><td>ITWeiHan</td><td>25</td><td>Taiwan</td></tr></tbody></table>";
-            var sourceData = new[] { new { Name = "ITWeiHan", Age = 25, Country = "Taiwan" } };
+            //IEnumrable
+            {
+                var sourceData = new[] { new { Name = "ITWeiHan", Age = 25, Country = "Taiwan" } };
 
-            var array = sourceData.ToArray().ToHtmlTable(tableAttributes:new {@class="SomeClass"});
-            Assert.AreEqual(array, expected);
+                var html = sourceData.ToArray().ToHtmlTable(tableAttributes: new { @class = "SomeClass" });
+                Assert.AreEqual(html, expected);
+            }
+
+            //DataTable
+            {
+                var sourceData = GetTestDataTable();
+
+                var html = sourceData.ToHtmlTable(tableAttributes: new { @class = "SomeClass" });
+                Assert.AreEqual(html, expected);
+            }
         }
 
         [TestMethod]
         public void AttributeTrTest()
         {
             var expected = @"<table><thead><tr class=""SomeClass"" ><th>Name</th><th>Age</th><th>Country</th></tr></thead><tbody><tr class=""SomeClass"" ><td>ITWeiHan</td><td>25</td><td>Taiwan</td></tr></tbody></table>";
-            var sourceData = new[] { new { Name = "ITWeiHan", Age = 25, Country = "Taiwan" } };
+            //IEnumrable
+            {
+                var sourceData = new[] { new { Name = "ITWeiHan", Age = 25, Country = "Taiwan" } };
 
-            var html = sourceData.ToArray().ToHtmlTable(trAttributes: new { @class = "SomeClass" });
-            Assert.AreEqual(html, expected);
+                var html = sourceData.ToArray().ToHtmlTable(trAttributes: new { @class = "SomeClass" });
+                Assert.AreEqual(html, expected);
+            }
+
+            //DataTable
+            {
+                var sourceData = GetTestDataTable();
+
+                var html = sourceData.ToHtmlTable(trAttributes: new { @class = "SomeClass" });
+                Assert.AreEqual(html, expected);
+            }
         }
 
         [TestMethod]
         public void AttributeTdTest()
         {
             var expected = @"<table><thead><tr><th>Name</th><th>Age</th><th>Country</th></tr></thead><tbody><tr><td class=""SomeClass"" >ITWeiHan</td><td class=""SomeClass"" >25</td><td class=""SomeClass"" >Taiwan</td></tr></tbody></table>";
-            var sourceData = new[] { new { Name = "ITWeiHan", Age = 25, Country = "Taiwan" } };
 
-            var html = sourceData.ToArray().ToHtmlTable(tdAttributes: new { @class = "SomeClass" });
-            Assert.AreEqual(html, expected);
+            //IEnumrable
+            {
+                var sourceData = new[] { new { Name = "ITWeiHan", Age = 25, Country = "Taiwan" } };
+
+                var html = sourceData.ToArray().ToHtmlTable(tdAttributes: new { @class = "SomeClass" });
+                Assert.AreEqual(html, expected);
+            }
+
+            //DataTable
+            {
+                var sourceData = GetTestDataTable();
+
+                var html = sourceData.ToHtmlTable(tdAttributes: new { @class = "SomeClass" });
+                Assert.AreEqual(html, expected);
+            }
         }
 
         [TestMethod]