+++ Changelog.md
@@ -1,3 +1,7 @@
+### version 1.1.5
+- [X] Fix Cache Key GetHashCode replaced by type.MetadataToken + type.Module.ModuleVersionId + prop.MetadataToken + prop.Module.ModuleVersionId
+- [X] Fix DataTable element attribute not work.
+
 ### version 1.1.4
 - [X] ToHtmlTableByProperties Add ExpressionCaches , it improve the efficiency after the cache created.
 - [X] Support Cunstom Table/TR/TD Attributes.